Ajout export CSV, statuts clients, README, corrections de securite

+++ supprimer.php
@@ -1,6 +1,11 @@
 <?php
 require "connexion.php";
 
+if (!isset($_POST['id']) || !ctype_digit($_POST['id'])) {
+    header("Location: liste.php");
+    exit;
+}
+
 $id = $_POST['id'];
 $stmt = $pdo->prepare("DELETE FROM contacts WHERE id = ?");
 $stmt->execute([$id]);
